note observed behaviour of makefile
Observations gleamed by 'make --debug=verbose' -> *.d files are only
imported if they exist when 'make' is run (from which we presumably they
are only necessary for deciding what to re-compile, not to perform a
clean-compile). TODO is a solution that creates (and imports) dependencies
before compilation, and whether there's any point-to/need-for such a
solution anyway. Consolidate makefiles.