ongoing
Add end 'summary'.
Example, a class made undeclareable (declaring a copy-ctor and only
copy-ctor results in no default/other-useable ctor).
Many ongoings: copying (what is at) 'char*', equivalency of
default-ctor/dtor to empty functions (with implicit calls to parent
ctors/dtors?), (what is to be learned from wikipedia example),
(efficency of) pass-by-value for assignment when using copy-and-swap,
exception safety alternatives to copy-and-swap.
TODO is (check?) whether self-assignment is a problem for any class in
this item.