extra/warnings-sanity-check
Taken from 'C++-examples' / bug from 'bench-write-read-text-files'
Example for '-Wconversion', bug due to incorrect use of
'long' instead of 'double' for an argument (which combined with subtle
double rounding errors can produce bugs which are easy to miss, as
there is no warning when passing a double as a long even with '-Wall').
Continue: a more reasonable set of maximalist warnings.